Vystrč (ne)fixtures do vlastních souborů
conftest.py, soubor automaticky načítaný Pytestem rozdělen na
fixtures.py a helpers.py. Zvyšuje se tím přehlednost – funkce jsou
jsou tam, kde by se daly očekávat podle názvu souboru. Conftest zní
přeci jen spíše jako nějaké nastavení.
